% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{affy.omni}
\alias{affy.omni}
\title{Mapping of Omni 2.5M probesets to Affy SNP6.0}
\format{
An annotated data.frame structure
\describe{
  \item{mergeid}{Unique snp identifier}
  \item{Probe_Set_ID}{Affy6 Probeset ID}
  \item{dbSNP_RS_ID}{dbSNP RS ID}
  \item{Chromosome}{Affy6 Chr}
  \item{Physical_Position}{Affy6 genomic position}
  \item{Strand}{Affy6 Strand}
  \item{Allele_A}{Affy6 Allele A}
  \item{Allele_B}{Affy6 Allele B}
  \item{Name}{Omni2.5 ID}
  \item{SNP}{Omni2.5 SNPs in X/Y format}
  \item{Chr}{Omni2.5 Chromosome}
  \item{MapInfo}{Omni2.5 hg19 genomic position}
  \item{RefStrand}{Omni2.5 Strand}
  \item{SNP_A}{Omni2.5 Allele A for RefStrand}
  \item{SNP_B}{Omni2.5 Allele B for RefStrand}
  \item{scSNP_A}{Omni2.5 strand-corrected Allele A}
  \item{scSNP_B}{Omni2.5 strand-corrected Allele B}
  \item{flip}{Value to flip Omni2.5 alleles if strands didnt match}
  ...
}
}
\usage{
data(affy.omni)
}
\description{
Mapping of Omni 2.5M probesets to Affy SNP6.0
}
\keyword{datasets}
