% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{annotateSegments}
\alias{annotateSegments}
\title{cnTools: annotate GRanges segments}
\usage{
annotateSegments(
  cn.data,
  genes,
  out.key = "SYMBOL",
  mart = NULL,
  use.mart = FALSE
)
}
\arguments{
\item{cn.data}{[Data.frame]: A dataframe that can be converted to GRanges object easily, or a granges object}

\item{genes}{[GRanges]: A GRanges object of genes with gene_ids housing annotation data. Easiest as the output from getGenes()}

\item{out.key}{Gene out.key, default is SYMBOL}

\item{mart}{[object]: A biomart object if you want to annotate missed genes with ensembl}

\item{use.mart}{[boolean]: If no mart is given, it will load in a biomart object}
}
\value{
Annotated GRanges object with gene ids for the input GRanges
}
\description{
cnTools: annotate GRanges segments
}
