% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{assemblePredDat}
\alias{assemblePredDat}
\title{assemblePredDat}
\usage{
assemblePredDat(D.vals, known.class = FALSE)
}
\arguments{
\item{D.vals}{}

\item{known.class}{}
}
\value{

}
\description{
Creates a melted data-structure designed for predictions,
much in the same vain as the training dataset: Sample X, sample 2, value
}
