% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{assignGrpIDs}
\alias{assignGrpIDs}
\title{pIDs}
\usage{
assignGrpIDs(mat, meta.df)
}
\arguments{
\item{mat}{}

\item{meta.df}{}
}
\description{
Creates a standardized set of group names for 
samples using a reference
}
