% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{balanceGrps}
\alias{balanceGrps}
\title{balanceGrps}
\usage{
balanceGrps(D.vals)
}
\arguments{
\item{D.vals}{}
}
\value{

}
\description{
Part of the process needed to balance the groups for a logistic
regression. It undersamples the non-matching group to be comparable with the
matching group
}
