% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{checkAgainst}
\alias{checkAgainst}
\title{checkAgainst}
\usage{
checkAgainst(mat)
}
\arguments{
\item{mat}{A matrix containing "cvclA" and "cvclB" columns for cellosaurus IDs of cell lines}
}
\value{
Character vector of OI (originating in), SS (synonymous), SI (sample from), 
and PCL (problematic)
}
\description{
Uses the assembled "prediction" matrix with CVCL ids attributed ot each
cell line to check whether the two cell line pairs are known to match in the 
Cellosaurus database
}
