% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkForConcordance.R
\name{checkForConcordance}
\alias{checkForConcordance}
\title{checkForConcordance}
\usage{
checkForConcordance(
  x.mat,
  metric = "euclidean",
  meta.dat = NULL,
  verbose = FALSE,
  return.matrix = FALSE,
  return.pred = TRUE,
  fill.na = TRUE,
  rm.gcsi = TRUE,
  sampleID = "Sample",
  meta.df
)
}
\arguments{
\item{x.mat}{Input matrix from compareVcf, or ref.mat}

\item{metric}{'euclidean', 'jaccard', or 'cor'}

\item{meta.dat}{Metadata, defaults to meta.df if left NULL (default:NULL)}

\item{verbose}{Verbose}

\item{return.matrix}{Return distance matrix (Default: FALSE)}

\item{return.pred}{Return predicted Match/Nonmatch (Default: TRUE)}

\item{fill.na}{Fill NA's in the input matrix with median value (Default=TRUE)}

\item{rm.gcsi}{Remove GNE samples from the input matrix (Default=TRUE)}

\item{sampleID}{Sample ID (Default='Sample')}

\item{meta.df}{Cell line metadata, accessible from CCLid::ccl_table}
}
\value{
A list containing Distance Matrix and list of Predictions (M/NM)
}
\description{
Checks for concordance between sample in first column of
input matrix and the rest of the reference matrix
}
