% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolute_ccl.R
\name{checkMse}
\alias{checkMse}
\title{checkMse}
\usage{
checkMse(M, M0)
}
\arguments{
\item{M}{Input matrix}

\item{M0}{Design matrix}
}
\description{
Finds the MSE for a NMF fit given a rank of 0 to 2.
If the MSE is lowest for rank 0, then this suggests that the 
design matrix is complete and perfectly explains the observed data.
If the MSE for a rank 1 or rank 2 is lowest, this suggests that
the design matrix partially explains the observed data.

Upon examining the H matrix, we can see the VERY rough proportion 
of the data in rank k that is unexplained from the design matrix.
}
