% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cin70}
\alias{cin70}
\title{Chromosomal instability (CIN) 70 genes from https://doi.org/10.1038/ng1861}
\format{
A data.frame structure
\describe{
  \item{Gene_old}{Published gene names from the paper, some are outdated}
  \item{Gene}{HUGO symbol ID updates for the old published genes}
  \item{core}{Whether the genes are core CIN genes (1) or not (0)}
  \item{ENS}{Ensemble ENS ID number}
  ...
}
}
\usage{
data(cin70)
}
\description{
Chromosomal instability (CIN) 70 genes from https://doi.org/10.1038/ng1861
}
\keyword{datasets}
