% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{combineSamples}
\alias{combineSamples}
\title{combineSamples}
\usage{
combineSamples(data.type, sample.mat, prop)
}
\arguments{
\item{data.type}{}

\item{sample.mat}{}

\item{prop}{}
}
\description{
Combines the BAF or Geno or two or samples given a set proportion
}
