% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{compareVcf}
\alias{compareVcf}
\title{compareVcf}
\usage{
compareVcf(
  vcfFile,
  var.dat,
  ref.mat,
  max.snps = 1e+06,
  ids = NULL,
  sampletype = "RNA",
  all.ids = NULL,
  ...
)
}
\arguments{
\item{vcfFile}{Path to VCF file to check against reference datasets}

\item{var.dat}{Variance data (list)}

\item{ref.mat}{Reference matrix (matrix)}

\item{max.snps}{Max number of SNPs to reduce}

\item{ids}{IDs}

\item{sampletype}{Strictly for labelling purposes}

\item{all.ids}{IDs to subset to}

\item{...}{Extra param}
}
\value{
Matrix: Containing reference matrix subsetted to common SNPs as 
the input VCF, as well as a left-joined VCF data
}
\description{
Checks a VCF file(s) against reference dataset
to look for similarity to any known cell lines
}
