% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{corWithDrug}
\alias{corWithDrug}
\title{corWithDrug}
\usage{
corWithDrug(dat.d, col.idx, title = "", text.thresh = 0.5)
}
\arguments{
\item{dat.d}{}

\item{col.idx}{}
}
\description{
IN DEVELOPMENT
}
