% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{demoSample}
\alias{demoSample}
\title{demoSample}
\usage{
demoSample(data.type = "BAF", sample.ord, demo.dat)
}
\arguments{
\item{data.type}{Must be either "BAF" (default) or "geno" (case-sensitive),}

\item{sample.ord}{}

\item{demo.dat}{}
}
\value{

}
\description{
Randomize a sample BAF or Geno giving probabilities
}
