% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{denis}
\alias{denis}
\title{denis}
\usage{
denis(vcf, refdir)
}
\arguments{
\item{vcf}{}

\item{refdir}{}
}
\value{
a denis object
}
\description{
Works with the webapp, passes in a vcf file and a
current directory for the reference matrix
}
