% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{.addCumPos}
\alias{.addCumPos}
\title{.addCumPos}
\usage{
.addCumPos(dat, ref, dat.type)
}
\arguments{
\item{dat}{Data containing $chrom, $pos, $loc.start, $loc.end}

\item{ref}{GRanges object containing $cum.start}

\item{dat.type}{Either "data" or "seg"}
}
\description{
.addCumPos
}
