% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.chkAndDownload}
\alias{.chkAndDownload}
\title{Check and download}
\usage{
.chkAndDownload(name, whichx, saveDir, verbose = FALSE)
}
\arguments{
\item{name}{Name of datatype to download and save to}

\item{whichx}{Row index of datatype to dowload}

\item{saveDir}{Directory that contains or will contain the .rda file}

\item{verbose}{Verbose}
}
\value{
Loads the .rda into .GlobalEnv
}
\description{
Checks for an existing .rda file of the input
data type. If it doesn't exist, it downloads from the
given CCLid::ccl_table data table URL and loads the .rda
file into memory
}
\keyword{internal}
