% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{.compSegs}
\alias{.compSegs}
\title{.compSegs}
\usage{
.compSegs(seg.obj, verbose = FALSE)
}
\arguments{
\item{seg.obj}{A seg obj}

\item{verbose}{Verbose (Default = FALSE)}
}
\value{
A list of seg objects
}
\description{
compare the difference between BAF profiles between pairwise
samples using a t-test of the raw probeset data
}
