% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{.estimateDrift}
\alias{.estimateDrift}
\title{.estimateDrift}
\usage{
.estimateDrift(seg.obj, ...)
}
\arguments{
\item{seg.obj}{DNAcopy::segment() object}

\item{z.cutoff}{Default set to 1:4;  Finds segments higher than those t-statistic away from 0}
}
\value{

}
\description{
Estimates genetic drift given an SD adjusted DNAcopy::segment()
object.  Estimates this based on a t-statistic
}
\examples{
.estimateDrift(seg.CNAo, z.cutoff=1:3)
}
