% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_drift.R
\name{.estimateZcutoff}
\alias{.estimateZcutoff}
\title{.estimateZcutoff}
\usage{
.estimateZcutoff(seg, data.type = "baf", verbose = FALSE)
}
\arguments{
\item{seg}{an output dataframe from a CNA object}

\item{data.type}{Data type, supports only 'baf' at this time}

\item{verbose}{Verbose, default = FALSE}
}
\value{
Returned seg with $t and $seg.diff
}
\description{
Creates a theoreticla framework for difference
between BAFs, and then compares the observed z-differences
against the theoretical diff to look for differences
}
\keyword{internal}
