% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{.filt}
\alias{.filt}
\title{.filt}
\usage{
.filt(vcf, min.depth = 5)
}
\arguments{
\item{vcf}{list from mapVcf2Affy}

\item{min.depth}{Default=5}
}
\value{
A filtered list from mapVcf2Affy
}
\description{
Removes probesets under a certain depth, especially useful
for RNAsequencing vcfs
}
