% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.fixGT}
\alias{.fixGT}
\title{.fixGT}
\usage{
.fixGT(vcf.gt, affy.gt, ret.idx = FALSE)
}
\arguments{
\item{vcf.gt}{VCF genotype (0,1,2)}

\item{affy.gt}{Affy6 genotype}

\item{ret.idx}{Returns the index of matching instead of genotype fix}
}
\description{
Switches the genotype to adjust for the reverse complement switch (0/0 -> 1/1)
}
\keyword{internal}
