% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.genMatchSnpDiff}
\alias{.genMatchSnpDiff}
\title{.genMatchSnpDiff}
\usage{
.genMatchSnpDiff(dr.nm, meta.df)
}
\arguments{
\item{dr.nm}{nonmatch SNP}

\item{meta.df}{Cell line metadata, accessible from CCLid::ccl_table}
}
\description{
generates a matrix of euclidean distance between probesets of 
sample pairs if the sample id's MATCH
}
\keyword{internal}
