% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellosaurus_accessors.R, R/utils.R
\name{.getDerivedFrom}
\alias{.getDerivedFrom}
\title{.getDerivedFrom}
\usage{
.getDerivedFrom(cvcl, melt.cells)

.getDerivedFrom(cvcl, melt.cells)
}
\arguments{
\item{cvcl}{CVCL id}

\item{melt.cells}{Melted cellosaurus dataframe, accessible from CCLid::ccl_table}

\item{meta.df}{Cell line metadata, accessible from CCLid::ccl_table}
}
\description{
.getDerivedFrom

.getDerivedFrom
}
\keyword{internal}
