% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{.getDrift}
\alias{.getDrift}
\title{.getDrift}
\usage{
.getDrift(i, idx = 1)
}
\arguments{
\item{i}{List object returned from bafDrift()}

\item{idx}{Index to return (1 = input compared to all matching cell lines, 2 = reference cell lines to all others)}
}
\value{

}
\description{
Returns the genomic fraction of drift for significant
regions (z>3)
}
