% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getSynonymous}
\alias{.getSynonymous}
\title{.getSynonymous}
\usage{
.getSynonymous(cvcl, melt.cells)
}
\arguments{
\item{cvcl}{CVCL id}

\item{melt.cells}{Melted cellosaurus dataframe, accessible from CCLid::ccl_table}
}
\description{
.getSynonymous
}
\keyword{internal}
