% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.getVariantFeatures}
\alias{.getVariantFeatures}
\title{.getVariantFeatures}
\usage{
.getVariantFeatures(ref.mat, bin.size = 1e+06, snp6.dat)
}
\arguments{
\item{ref.mat}{Reference matrix of Samples by Probesets}

\item{bin.size}{Bin size, Default=1e6}

\item{snp6.dat}{SNP6 probeset genomic position, accessible from CCLid::ccl_table}
}
\description{
Calculates the features/probesets with the most variance from a 
given matrix (rownames).  Returns equally spaced Variant Probesets
}
\keyword{internal}
