% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.jsonToGr}
\alias{.jsonToGr}
\title{.jsonToGr}
\usage{
.jsonToGr(json, from.file = TRUE)
}
\arguments{
\item{json}{Json VCF format - OUTDATED}

\item{from.file}{Boolean to say whether it comes from a file or passed in object}
}
\description{
converts a JSON object into a VariantAnnotation object
}
\keyword{internal}
