% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{.normBAF}
\alias{.normBAF}
\title{.normBAF}
\usage{
.normBAF(x, lower = T)
}
\arguments{
\item{x}{BAF vector}

\item{lower}{if TRUE, 0-0.5 range, ELSE, 0.5-1 range}
}
\description{
rescales the BAF from a 0-1 scale to a 0-0.5
}
