% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPos.R
\name{.overlapProbeset}
\alias{.overlapProbeset}
\title{.overlapProbeset}
\usage{
.overlapProbeset(ref.ids, comp.ids)
}
\arguments{
\item{ref.ids}{Reference probeset IDs to match order to}

\item{comp.ids}{IDs of the given input, comparator}
}
\description{
Overlaps probesets and orders based on chr position
}
