% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{.revComp}
\alias{.revComp}
\title{.revComp}
\usage{
.revComp(vcf.gt, ret.idx = FALSE)
}
\arguments{
\item{vcf.gt}{VariantAnnotation ob ject}

\item{ret.idx}{Boolean, if True, does not reverse complement any alleles}
}
\description{
Reverse complements the Alleles
}
