% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{.vcf2AffyGT}
\alias{.vcf2AffyGT}
\title{.vcf2AffyGT}
\usage{
.vcf2AffyGT(vcf.gt)
}
\arguments{
\item{vcf.gt}{VariantAnnotation object where the $GT column contains 1/1, 0/1, etc...}
}
\value{

}
\description{
Maps Genotypes called from VCFs (0/0, 0/1, 1/0, 1/1) to the Affymetrix
standard of 0, 1, 2 genotypes
}
\examples{
.vcf2AffyGT(vcf.affy.gr)
}
