% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_genotype.R
\name{.zval}
\alias{.zval}
\title{z-statistic}
\usage{
.zval(dat, p = FALSE)
}
\arguments{
\item{dat}{vector of numbers to calculate z stat from}

\item{p}{returns p-value instead of z-statistic}
}
\description{
z-statistic
}
