% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{downloadRefCCL}
\alias{downloadRefCCL}
\title{downloadRefCCL}
\usage{
downloadRefCCL(
  name,
  saveDir = file.path(".", "CCLid"),
  refFileName = NULL,
  verbose = FALSE,
  bin.size = NULL
)
}
\arguments{
\item{name}{Currently only supports "BAF"}

\item{saveDir}{Directory containign bigmemory .desc/.bin folder, or where to save new download}

\item{refFileName}{File name to save, or pre-existing rds or .bin file}

\item{verbose}{Verbose (Default=FALSE)}

\item{bin.size}{Genomic size to chunk the SNPs into}
}
\description{
Downloads the precomputed RDS data structures for reference
}
