% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{driftOverlap}
\alias{driftOverlap}
\title{driftOverlap}
\usage{
driftOverlap(seg, ref.ds = NULL, alt.ds = NULL)
}
\arguments{
\item{seg}{}

\item{ref.ds}{}

\item{alt.ds}{}
}
\value{

}
\description{
driftOverlap
}
