% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{driftOverlapMetric}
\alias{driftOverlapMetric}
\title{driftOverlapMetric}
\usage{
driftOverlapMetric(gr.baf, gr.cn, cell.ids, ov.frac = seq(0, 1, by = 0.01))
}
\arguments{
\item{gr.baf}{GRangesList of cell lines an their BAF-drifted regions}

\item{gr.cn}{GRangesList of cell lines an their CN-drifted regions}

\item{cell.ids}{Vector of cell line names to check drift of}

\item{ov.frac}{overlap fraction cutoff: (Default: seq(0, 1, by=0.01))}
}
\value{
List containing the following elements:
"model" = non-linear least-square model fitted to concordance and sensitvity
"saturation"=Matrix of total saturation and concordance cutoff
"sens"=Matrix of sensitivity value for each concordance cutoff
"dat"=Matrix of concordance cutoff by cell lines
}
\description{
Calculate the amount of overlap between the CN and BAF inferred drift
GRanges objects
}
