% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPos.R
\name{findCclPairs}
\alias{findCclPairs}
\title{findCclPairs}
\usage{
findCclPairs(meta.df, dr.nm)
}
\arguments{
\item{meta.df}{}

\item{dr.nm}{matrix where column names are cell lines/datasets}
}
\value{

}
\description{
Finds indices for cell lines that are found
in multiple datasets
}
