% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{formatRefMat}
\alias{formatRefMat}
\title{Title}
\usage{
formatRefMat(
  name,
  ref.mat,
  analysis,
  varFileName = NULL,
  saveDir = file.path(".", "CCLid"),
  bin.size = 1e+06,
  just.var = FALSE
)
}
\arguments{
\item{ref.mat}{}

\item{analysis}{}
}
\value{

}
\description{
Title
}
