% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullpull.R
\name{fullpull}
\alias{fullpull}
\title{fullpull}
\usage{
fullpull(cvcl)
}
\arguments{
\item{cvcl}{Cellosaurus CVCL id (e.g. CVCL_1384)}
}
\value{
Dataframe of all synonymous (SS), similar origins (OI), MSI positive, or problematic
lines associated with the CVCL
}
\description{
Gets the synonymous cell lines and cell line that they derived from,
as well as any indication of MSI or probelmatic cell line (e.g. contamination)
}
\examples{
cvcl <- getCVCL('Hela')
fullpull(cvcl)
}
