% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{genDemoData}
\alias{genDemoData}
\title{genDemoData}
\usage{
genDemoData(data.type = "BAF", n.pop = 10, ...)
}
\arguments{
\item{data.type}{}

\item{n.pop}{}

\item{...}{}
}
\value{

}
\description{
Generates random demo data to play with
}
