% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{genErrBp}
\alias{genErrBp}
\title{genErrBp}
\usage{
genErrBp(p.m.nm)
}
\arguments{
\item{p.m.nm}{output matrix of SplitToMNM}
}
\value{
Dataframe of errs and pcls
}
\description{
Takes the output of splitToMNM() and reduces it to a dataframe
of Errors and PCLs
}
