% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{getCinScore}
\alias{getCinScore}
\title{getCinScore}
\usage{
getCinScore(psets, cin.metric = "sum")
}
\arguments{
\item{psets}{}

\item{cin.metric}{}
}
\value{
a CIN list
}
\description{
getCinScore
}
