% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCVCL.R
\name{getCVCL}
\alias{getCVCL}
\title{getCVCL}
\usage{
getCVCL(cellid, prioritize.datasets = TRUE, melt.cells)
}
\arguments{
\item{cellid}{character: cell line name (e.g. MCF-7)}

\item{prioritize.datasets}{default=TRUE}

\item{melt.cells}{Melted cellosaurus dataframe, accessible from CCLid::ccl_table}
}
\value{
CVCL_ style character name
}
\description{
get the Cellosaurus CVCL_ id for a given
cell line name
}
