% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{getGeneExpr}
\alias{getGeneExpr}
\title{getGeneExpr}
\usage{
getGeneExpr(psets, gene.id, in.key = "SYMBOL")
}
\arguments{
\item{psets}{PSets from pharmacoGX}

\item{gene.id}{Gene ID to map to Ensembl IDs}

\item{in.key}{for mapping gene IDs, type of Gene (Default=SYMBOL)}
}
\description{
getGeneExpr
}
