% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{getVcfDrifts}
\alias{getVcfDrifts}
\title{getVcfDrifts}
\usage{
getVcfDrifts(vcfFile, ref.dat, rna.meta.df, ...)
}
\arguments{
\item{vcfFile}{path to vcf file}

\item{rna.meta.df}{Meta file linking RNA files to cell names}
}
\value{
A list containing the fraction of genome drifted,
as well the significantly drifted regions CNAo
}
\description{
Check the drift of a given VCF file and all matching
cell line names based on the meta-data matched IDs
}
