% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gne.meta}
\alias{gne.meta}
\title{Metadata for gCSI dataset}
\format{
A data.frame structure
\describe{
  ...
}
}
\usage{
data(gne.meta)
}
\description{
Metadata for gCSI dataset
}
\keyword{datasets}
