% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{mapVariantFeat}
\alias{mapVariantFeat}
\title{mapVariantFeat}
\usage{
mapVariantFeat(vcf.map, var.dat)
}
\arguments{
\item{vcf.map}{An object returned by mapVcf2Aff()}

\item{var.dat}{A list returned by formatRefMat()$var}

\item{slow.method}{}
}
\value{

}
\description{
Selects only the features with the highest variance
per bin
}
