% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{mapVcf2Affy}
\alias{mapVcf2Affy}
\title{mapVcf2Affy}
\usage{
mapVcf2Affy(vcfFile)
}
\arguments{
\item{vcfFile}{Absolute path to VCF file}
}
\value{
List composed of two dataframes, BAF and GT
}
\description{
Takes a VCF files and reads it into VariantAnnotation package to overlap with
GRanges probeset dataset of snp6.dat.  It willc onvert the 0/0, 0/1, etc.. genotypes to 0,1,2.
Finally, it will reutnr the BAFs and normalized BAFs for each of the overlapping probesets.
}
\examples{
vcfFile <- "/mnt/work1/users/pughlab/projects/cancer_cell_lines/denis_id/mutect_GDSC/EGAR00001252191_13305_1/EGAR00001252191_13305_1.vcf"
mapVcf2Affy(vcfFile)
}
