% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meta.df}
\alias{meta.df}
\title{Mapping of cell IDs to filenames}
\format{A data.frame structure
\describe{
  \item{ID}{Unique cell name}
  \item{GDSC}{Filename ID for GDSC data}
  \item{CCLE}{Filename ID for CCLE data}
  ...
}}
\usage{
data(meta.df)
}
\description{
Mapping of cell IDs to filenames
}
\keyword{datasets}
