% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{multiDriftPlot}
\alias{multiDriftPlot}
\title{multiDriftPlot OUTDATED}
\usage{
multiDriftPlot(seg, chr.size.gr = NULL, ref.ds = NULL, alt.ds = NULL)
}
\arguments{
\item{seg}{Seg data}

\item{chr.size.gr}{NULL}

\item{ref.ds}{NULL}

\item{alt.ds}{NULL}
}
\description{
multiDriftPlot OUTDATED
}
