% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.CCLid}
\alias{plot.CCLid}
\title{plot.CCLid
plot() function for CCLid adjusted DNAcopy segment objects}
\usage{
\method{plot}{CCLid}(
  obj,
  sample.size = 600,
  low.sig.alpha = 0.01,
  hi.sig.alpha = 0.2,
  add.chr.sep = TRUE,
  atype = "sd",
  add.points = TRUE,
  min.z = 3,
  ...
)
}
\arguments{
\item{obj}{CCLid object returned from bafDrift()}

\item{sample.size}{Number of points to plot to allow for quicker plotting}

\item{low.sig.alpha}{low-confident drift regions alpha value for red}

\item{hi.sig.alpha}{high-confident drift regions alpha value for red}

\item{add.chr.sep}{Adds lines to delimit chromosomes}

\item{atype}{default=sd}

\item{add.points}{Adds points to the segments}

\item{min.z}{minimum z to report a drift change}
}
\description{
plot.CCLid
plot() function for CCLid adjusted DNAcopy segment objects
}
