% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.CCLidDRIFT}
\alias{plot.CCLidDRIFT}
\title{plot.CCLidDRIFT
plot() function for CCLid adjusted DNAcopy segment objects}
\usage{
plot.CCLidDRIFT(
  obj,
  sample.size = 50,
  low.sig.alpha = 0.01,
  hi.sig.alpha = 0.2
)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
plot.CCLidDRIFT
plot() function for CCLid adjusted DNAcopy segment objects
}
