% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.CCLidDRIFTMulti}
\alias{plot.CCLidDRIFTMulti}
\title{plot.CCLidDRIFTMulti}
\usage{
plot.CCLidDRIFTMulti(seg, chr.size.gr = NULL, ref.ds = NULL, alt.ds = NULL)
}
\arguments{
\item{seg}{}

\item{chr.size.gr}{}

\item{ref.ds}{}

\item{alt.ds}{}
}
\value{

}
\description{
plot.CCLidDRIFTMulti
}
