% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.multiObj}
\alias{plot.multiObj}
\title{plot.multiObj}
\usage{
plot.multiObj(OBJ, add.chr.sep = TRUE, min.z = 2)
}
\arguments{
\item{OBJ}{A list of CCLid Objects}

\item{add.chr.sep}{}

\item{min.z}{}
}
\description{
plot.multiObj
}
