% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.SimpleCCLid}
\alias{plot.SimpleCCLid}
\title{plot.SimpleCCLid}
\usage{
plot.SimpleCCLid(obj, add.chr.sep = TRUE, min.z = 2)
}
\arguments{
\item{obj}{}

\item{add.chr.sep}{}

\item{min.z}{}
}
\description{
plot.SimpleCCLid
}
