% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{plotFracDrift}
\alias{plotFracDrift}
\title{plotFracDrift}
\usage{
plotFracDrift(summ.frac)
}
\arguments{
\item{summ.frac}{summ frac list to plot, contains $baf and $cn}
}
\description{
plotFracDrift
}
