% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{plotHist}
\alias{plotHist}
\title{plotHist}
\usage{
plotHist(D)
}
\arguments{
\item{D}{Returns from predictions, containins $M and $NM as elements of the list}
}
\description{
Plots the histogram for matching/non-matching dataset
}
