% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{readinRnaFileMapping}
\alias{readinRnaFileMapping}
\title{readinRnaFileMapping}
\usage{
readinRnaFileMapping()
}
\value{

}
\description{
Map the RNA files to the SNP files
using hardcoded metadata
}
