% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rna.meta.df}
\alias{rna.meta.df}
\title{Metadata mapping together all pharmacogenomic datasets and their RNA IDs}
\format{
A data.frame structure
\describe{
  ...
}
}
\usage{
data(rna.meta.df)
}
\description{
Metadata mapping together all pharmacogenomic datasets and their RNA IDs
}
\keyword{datasets}
