% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{sigDiffBaf}
\alias{sigDiffBaf}
\title{sigDiffBaf}
\usage{
sigDiffBaf(each.sample)
}
\arguments{
\item{each.sample}{Each sample}
}
\value{
GRanges object of significant BAF drift (z > 3)
for a cna.obj returned by bafDrift$cna.obj
}
\description{
sigDiffBaf
}
