% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{similarityMatrix}
\alias{similarityMatrix}
\title{similarityMatrix}
\usage{
similarityMatrix(mat, method)
}
\arguments{
\item{mat}{Matrix of BAFs}

\item{method}{Either 'cor', 'jaccard', or 'euclidean'}
}
\description{
Creates a similarity/concordance matrix given a
SNP x Sample matrix and a method of concordance
}
