% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp6.dat}
\alias{snp6.dat}
\title{BED file of SNP6 Probeset IDs and genomic loci in hg19}
\format{
A list containing 3 elements, each containing 'ProbesetID, chrom, pos, strand'
\describe{
  \item{SNP}{SNP probeset data-frame containing 933422 rows x 4 cols}
  \item{CN}{CN probeset data-frame containing 945615 rows x 4 cols}
  \item{All}{SNP and CN probeset data-frame containing 1879037 rows x 4 cols}
  ...
}
}
\source{
\url{http://www.affymetrix.com/Auth/analysis/downloads/lf/genotyping/GenomeWideSNP_6/GenomeWideSNP_6.hg19.bed.zip}
}
\usage{
data(snp6.dat)
}
\description{
BED file of SNP6 Probeset IDs and genomic loci in hg19
}
\keyword{datasets}
