% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitConcordanceVals.R
\name{splitConcordanceVals}
\alias{splitConcordanceVals}
\title{splitConcordanceVals}
\usage{
splitConcordanceVals(dm, meta.df)
}
\arguments{
\item{dm}{dataframe melt}

\item{meta.df}{metadata for the melted dataframe}
}
\description{
Melts the data structure into Sample 1, Sample 2, concordance
and separates into Match/nonmatch if a reference metadata is given
}
