% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{splitToMNM}
\alias{splitToMNM}
\title{splitToMNM}
\usage{
splitToMNM(p)
}
\arguments{
\item{p}{prediction data frame}
}
\value{
A subset of p
}
\description{
Reduces the prediction dataframe to samples with a predicted
Match and a ground truth of non-match (based on anontations)
}
