% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{summarizeFracDrift}
\alias{summarizeFracDrift}
\title{summarizeFracDrift}
\usage{
summarizeFracDrift(cn.drifts, cn.z, baf.drifts, baf.z, include.id = FALSE)
}
\arguments{
\item{cn.drifts}{}

\item{cn.z}{}

\item{baf.drifts}{}

\item{baf.z}{}
}
\value{
list of BAF and CN summary frac
}
\description{
Summarizes the $frac from the baf.drifts and cn.drifts
given a cutoff for cn or baf
}
