% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{annoDemoMat}
\alias{annoDemoMat}
\title{annoDemoMat}
\usage{
annoDemoMat(sample.ord, demo.mat)
}
\arguments{
\item{sample.ord}{}

\item{demo.mat}{}
}
\value{

}
\description{
Annotate the demo SNP x Sample matrix to generate meta.df
}
