% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{assemblePredDat}
\alias{assemblePredDat}
\title{assemblePredDat}
\usage{
assemblePredDat(D.vals, known.class = FALSE)
}
\arguments{
\item{D.vals}{Group values (M/NM)}

\item{known.class}{Default = FALSE, hardcodes a M/NM for demo data}
}
\description{
Creates a melted data-structure designed for predictions,
much in the same vain as the training dataset: Sample X, sample 2, value
}
