% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{availableRefCCL}
\alias{availableRefCCL}
\title{availableRefCCL}
\usage{
availableRefCCL(
  saveDir = file.path(".", "CCLid"),
  tableDir = NULL,
  myfn = "downloadTable.csv",
  verbose = TRUE
)
}
\arguments{
\item{saveDir}{Directory to store Ref files}

\item{tableDir}{Defaault:NULL, uses CCLids built in table detailing external
data location}

\item{myfn}{Default: "downloadTable.csv"}

\item{verbose}{Default: TRUE}
}
\description{
table describing the different datasets for download
}
