% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bafDrift.R
\name{bafDrift}
\alias{bafDrift}
\title{bafDrift}
\usage{
bafDrift(
  sample.mat,
  debug = FALSE,
  centering = "none",
  norm.baf = TRUE,
  hom.filt.val = 0.07,
  snp6.dat,
  ...
)
}
\arguments{
\item{sample.mat}{Input sample matrix}

\item{debug}{should be set to FALSE and only changed when debugging}

\item{centering}{Centering the metric on median, none, or mean}

\item{norm.baf}{Boolean to normalize BAF (Default =TRUE)}

\item{hom.filt.val}{Homozygous filtering threshold (Default = 0.07)}

\item{snp6.dat}{SNP6 probeset genomic position, accessible from CCLid::ccl_table}

\item{...}{Extra param}
}
\description{
Calcualtes the amount of genetic drift in a sample
}
