% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{bafDrift}
\alias{bafDrift}
\title{bafDrift}
\usage{
bafDrift(sample.mat, debug = FALSE, centering = "none", ...)
}
\arguments{
\item{sample.mat}{}

\item{debug}{should be set to FALSE and only changed when debugging}
}
\value{

}
\description{
bafDrift
}
