% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{checkForConcordance}
\alias{checkForConcordance}
\title{checkForConcordance}
\usage{
checkForConcordance(x.mat, metric = "euclidean", meta.dat = NULL,
  verbose = FALSE, return.matrix = FALSE, return.pred = TRUE,
  fill.na = TRUE, rm.gcsi = TRUE, sampleID = "Sample")
}
\arguments{
\item{x.mat}{Input matrix from compareVcf, or ref.mat}

\item{metric}{'euclidean', 'jaccard', or 'cor'}

\item{meta.dat}{Metadata, defaults to meta.df if left NULL (default:NULL)}

\item{verbose}{}

\item{return.matrix}{Return distance matrix (Default: FALSE)}

\item{return.pred}{Return predicted Match/Nonmatch (Default: TRUE)}

\item{fill.na}{Fill NA's in the input matrix with median value (Default=TRUE)}

\item{rm.gcsi}{Remove GNE samples from the input matrix (Default=TRUE)}

\item{sampleID}{Sample ID (Default='Sample')}
}
\value{
A list containing Distance Matrix and list of Predictions (M/NM)
}
\description{
Checks for concordance between sample in first column of
input matrix and the rest of the reference matrix
}
\examples{
  ref.dat <- CCLid::loadRef(PDIR, 'baf', bin.size=5e5)
  ref.mat <- ref.dat$ref
  var.dat <- ref.dat$var
  vcfFile <- '/mnt/work1/users/home2/quever/xfer/A549.sample_id.vcf' ## A549 WES
  vcf.mat <- compareVcf(vcfFile, var.dat=ref.dat$var, ref.mat=ref.dat$ref)
  checkForConcordance(vcf.mat)
}
