% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{compareVcf}
\alias{compareVcf}
\title{compareVcf}
\usage{
compareVcf(
  vcfFile,
  var.dat,
  ref.mat,
  max.snps = 1e+06,
  ids = NULL,
  sampletype = "RNA",
  ...
)
}
\arguments{
\item{vcfFile}{Path to VCF file to check against reference datasets}

\item{var.dat}{Variance data (list)}

\item{ref.mat}{Reference matrix (matrix)}

\item{max.snps}{}

\item{ids}{}

\item{sampletype}{}

\item{...}{}
}
\value{
Matrix: Containing reference matrix subsetted to common SNPs as 
the input VCF, as well as a left-joined VCF data
}
\description{
Checks a VCF file(s) against reference dataset
to look for similarity to any known cell lines
}
\examples{
  ref.dat <- CCLid::loadRef(PDIR, 'baf', bin.size=5e5)
  ref.mat <- ref.dat$ref
  var.dat <- ref.dat$var
  vcfFile <- '/mnt/work1/users/home2/quever/xfer/A549.sample_id.vcf' ## A549 WES
  vcf.mat <- compareVcf(vcfFile, var.dat=ref.dat$var, ref.mat=ref.dat$ref)
}
