% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{corWithDrug}
\alias{corWithDrug}
\title{corWithDrug}
\usage{
corWithDrug(dat.d, col.idx, title = "", text.thresh = 0.5, genes = NULL)
}
\arguments{
\item{dat.d}{Data D containg $tCIN}

\item{col.idx}{column index}

\item{title}{titleof plot}

\item{text.thresh}{text.threshold of 0.50}

\item{genes}{List of genes}
}
\description{
IN DEVELOPMENT
}
