% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{.addSegSd}
\alias{.addSegSd}
\title{addSegSd}
\usage{
.addSegSd(seg.obj)
}
\arguments{
\item{seg.obj}{an object returned from DNAcopy::segment()}
}
\description{
Adds segment SD to CBS segment objects
}
