% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{.estimateZcutoff}
\alias{.estimateZcutoff}
\title{.estimateZcutoff}
\usage{
.estimateZcutoff(seg, data.type = "baf")
}
\arguments{
\item{seg}{an output dataframe from a CNA object}
}
\value{
Returned seg with $t and $seg.diff
}
\description{
Creates a theoreticla framework for difference
between BAFs, and then compares the observed z-differences
against the theoretical diff to look for differences
}
