% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_probesets.R
\name{.fixGT}
\alias{.fixGT}
\title{.fixGT}
\usage{
.fixGT(vcf.gt, affy.gt, ret.idx = FALSE)
}
\arguments{
\item{vcf.gt}{}

\item{affy.gt}{}
}
\value{

}
\description{
Switches the genotype to adjust for the reverse complement switch (0/0 -> 1/1)
}
