% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{.genMatchSnpDiff}
\alias{.genMatchSnpDiff}
\title{.genMatchSnpDiff}
\usage{
.genMatchSnpDiff(dr.nm)
}
\arguments{
\item{dr.nm}{nonmatch SNP}
}
\description{
generates a matrix of euclidean distance between probesets of 
sample pairs if the sample id's MATCH
}
