% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{.genNonmatchSnpDiff}
\alias{.genNonmatchSnpDiff}
\title{.genNonmatchSnpDiff}
\usage{
.genNonmatchSnpDiff(col.len, dr.nm)
}
\arguments{
\item{col.len}{Col length}

\item{dr.nm}{Nonmatch SNP}
}
\description{
generates a matrix of euclidean distance between probesets of 
sample pairs if the sample id's dont' match (NON-MATCH)
}
