% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPos.R
\name{.overlapProbeset}
\alias{.overlapProbeset}
\title{.overlapProbeset}
\usage{
.overlapProbeset(ref.ids, comp.ids)
}
\arguments{
\item{ref.ids}{}

\item{comp.ids}{}
}
\value{

}
\description{
Overlaps probesets and orders based on chr position
}
