% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.overlapProbeset}
\alias{.overlapProbeset}
\title{.overlapProbeset}
\usage{
.overlapProbeset(ref.ids, comp.ids, snp6.dat)
}
\arguments{
\item{ref.ids}{Reference probeset IDs to match order to}

\item{comp.ids}{IDs of the given input, comparator}

\item{snp6.dat}{SNP6 probeset genomic position, accessible from CCLid::ccl_table}
}
\description{
Overlaps probesets and orders based on chr position
}
\keyword{internal}
