% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadRefCCL.R
\name{downloadRefCCL}
\alias{downloadRefCCL}
\title{downloadRefCCL}
\usage{
downloadRefCCL(name, saveDir = file.path(".", "CCLid"), verbose = FALSE)
}
\arguments{
\item{name}{Currently only supports "BAF"}

\item{saveDir}{Directory containign bigmemory .desc/.bin folder, or where to save new download}

\item{verbose}{Verbose (Default=FALSE)}
}
\description{
Downloads the precomputed RDS data structures for reference
}
