% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_ref.R
\name{downloadRefCCL}
\alias{downloadRefCCL}
\title{downloadRefCCL}
\usage{
downloadRefCCL(
  name,
  saveDir = file.path(".", "CCLid"),
  refFileName = NULL,
  verbose = FALSE,
  bin.size = NULL
)
}
\arguments{
\item{name}{}

\item{saveDir}{}

\item{refFileName}{}

\item{verbose}{}
}
\value{

}
\description{
Downloads the precomputed RDS data structures for reference
}
\examples{
downloadRefCCL(name="BAF")
}
