% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/driftOverlap.R
\name{driftOverlap}
\alias{driftOverlap}
\title{driftOverlap}
\usage{
driftOverlap(seg, ref.ds = NULL, alt.ds = NULL)
}
\arguments{
\item{seg}{Seg data}

\item{ref.ds}{Reference dataset}

\item{alt.ds}{Alternate dataset}
}
\description{
driftOverlap
}
