% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatRefMat.R
\name{formatRefMat}
\alias{formatRefMat}
\title{formatRefMat}
\usage{
formatRefMat(
  name,
  ref.mat,
  analysis = "baf",
  varFileName = NULL,
  saveDir = file.path(".", "CCLid"),
  bin.size = 1e+06,
  just.var = FALSE,
  fill.na = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{name}{Name to be applied to the returned object}

\item{ref.mat}{A reference matrix of all pharmacogenomic reference samples and their BAFs,
typically obtained from the downloadRefCCL() function}

\item{analysis}{Only 'baf' is implemented at this point}

\item{varFileName}{RDS containing the variant SNP information}

\item{saveDir}{Directory to save/load the variant RDS file to}

\item{bin.size}{Default is set to 5e5, a variant file must be created for this bin size}

\item{just.var}{Just runs the variant SNP part of the script, skips subsetting}

\item{fill.na}{Fills NA with median (Default=FALSE)}

\item{verbose}{Verbose}
}
\value{
Returns a list object:
'ref' = matrix of SNPs by samples for least variant SNPs
'var' = list of each 'bin size' and the SNPs and their BAF that populate it
}
\description{
Pre-processes the input data into a matrix of BAF values and 
the least/most variant SNPs within that cohort
}
