% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_data.R
\name{genDemoData}
\alias{genDemoData}
\title{genDemoData}
\usage{
genDemoData(data.type = "BAF", n.pop = 10, ...)
}
\arguments{
\item{data.type}{Must be "BAF" (default) or "geno" (case-sensitive)}

\item{n.pop}{Number of samples to simulate}

\item{...}{Pass in}
}
\value{
A list object
 matrix = A matrix of 'geno' or 'BAFs' for n.pop samples
 meta = a matrix of all by all sample comparisons
 prob = a probability matrix used to generate each sample
}
\description{
Generates random demo data to play with
}
