% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{getBafDrifts}
\alias{getBafDrifts}
\title{getBafDrifts}
\usage{
getBafDrifts(cl.pairs, x.mat, ref.ds = NULL, alt.ds = NULL, ...)
}
\arguments{
\item{cl.pairs}{Vector of indices for cell line pairs in x.mat}

\item{x.mat}{Input matrix containing probeset BAF data}

\item{ref.ds}{Reference dataset (e.g. CCLE)}

\item{alt.ds}{Comparing dataset (e.g. GDSC)}
}
\value{
Drift object
}
\description{
Gets the drift GRanges and genomic fraction of drift
for any pair of cell lines from different datasets given
}
