% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper_functions.R
\name{getCNDrifts}
\alias{getCNDrifts}
\title{getCnDrifts}
\usage{
getCNDrifts(
  ref.l2r,
  alt.l2r,
  fdat,
  seg.id,
  raw.id,
  cell.ids,
  verbose = TRUE,
  meta.df,
  ...
)
}
\arguments{
\item{ref.l2r}{Matrix of L2R data of genomic bins by samples for reference dataset}

\item{alt.l2r}{Matrix of L2R data of genomic bins by samples for comparison dataset}

\item{fdat}{segment data}

\item{seg.id}{element ID of ref.l2r that specifies the segment L2Rs}

\item{raw.id}{element ID of ref.l2r that specifies the raw probeset L2Rs}

\item{cell.ids}{All cell line IDs to compare drift between}

\item{verbose}{Verbose}

\item{meta.df}{Cell line metadata, accessible from CCLid::ccl_table}

\item{...}{Extra param}
}
\value{
CN drift object
}
\description{
Gets the drift GRanges and genomic fraction of drift
for any pair of cell lines from different datasets given L2R data
}
