% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{loadRef}
\alias{loadRef}
\title{loadRef}
\usage{
loadRef(pdir = NULL, analysis = "baf", ...)
}
\arguments{
\item{pdir}{Path to directory containing or to download reference data}

\item{analysis}{Only support BAF at the moment}

\item{...}{bin.size to group variance data}
}
\value{
List containing "ref"=reference BAF matrix
and "var"=list containing variance for bin-sizes
}
\description{
Wrapper to load in all the reference matrices
and annotate them with groups, as well as calculate variance
}
\examples{
PDIR <- "/mnt/work1/users/pughlab/projects/cancer_cell_lines/CCL_paper/CCLid/CCLid"
loadRef(PDIR, 'baf', bin.size=5e5)
}
