% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{loadRef}
\alias{loadRef}
\title{loadRef}
\usage{
loadRef(
  PDIR = NULL,
  analysis = "baf",
  rm.gne = FALSE,
  bin.size = 1e+06,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{PDIR}{Directory for download data}

\item{analysis}{Only support BAF at the moment}

\item{rm.gne}{Remove gCSI from the analysis (default=FALSE)}

\item{bin.size}{Bin size (default=1e6)}

\item{verbose}{Verbose}

\item{...}{Extra param}
}
\value{
List containing "ref"=reference BAF matrix
and "var"=list containing variance for bin-sizes
}
\description{
Wrapper to load in all the reference matrices
and annotate them with groups, as well as calculate variance
}
