% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{mkPredictions}
\alias{mkPredictions}
\title{mkPredictions}
\usage{
mkPredictions(pred, models)
}
\arguments{
\item{pred}{Prediction data}

\item{models}{logistic reression model}
}
\description{
Uses a trained model to predict on a pred data structure
}
