% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPos.R
\name{overlapPos}
\alias{overlapPos}
\title{overlapPos}
\usage{
overlapPos(comp, ref, mapping = "probeset")
}
\arguments{
\item{comp}{}

\item{ref}{}

\item{mapping}{}
}
\value{

}
\description{
parent function that handles how overlaps are done between COMParing 
group and REFerence datasets
}
