% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapPos.R
\name{overlapPos}
\alias{overlapPos}
\title{overlapPos}
\usage{
overlapPos(comp, ref, mapping = "probeset")
}
\arguments{
\item{comp}{Input comparison data with a $Probe_Set_ID column}

\item{ref}{Reference probeset order where rownames are Probe_Set_IDs}

\item{mapping}{If "probeset", will using $Probe_Set_ID from comp dataframe.
Else, it will use the rownames of ref as the probe set IDs}
}
\description{
parent function that handles how overlaps are done between COMParing 
group and REFerence datasets
}
