% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_drift.R
\name{plot.CCLid}
\alias{plot.CCLid}
\title{plot.CCLid
plot() function for CCLid adjusted DNAcopy segment objects}
\usage{
plot.CCLid(
  obj,
  sample.size = 50,
  low.sig.alpha = 0.01,
  hi.sig.alpha = 0.2,
  add.chr.sep = TRUE,
  atype = "sd",
  add.points = TRUE
)
}
\arguments{
\item{obj}{}
}
\value{

}
\description{
plot.CCLid
plot() function for CCLid adjusted DNAcopy segment objects
}
