% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drift_calc.R
\name{segmentDrift}
\alias{segmentDrift}
\title{segmentDrift}
\usage{
segmentDrift(segmenter = "PCF", fdat, D, kmin = 5, rm.homo = FALSE)
}
\arguments{
\item{segmenter}{either CBS or PCF (Default: PCF)}

\item{fdat}{Genomic data-frame where first two columns are "chrom" and "pos"}

\item{D}{Matrix corresponding to fdat containing distances to plot, samples in 
columns, rows are the genomic pos}

\item{kmin}{Minimum number of SNPs to consider (default = 5)}

\item{rm.homo}{Remove homozygous SNPs (Default = FALSE)}
}
\value{
CNA object
}
\description{
Segments a matrix (D) given a set of genomic coordinates
stored in fdat using either CBS or PCF
}
