% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pred_cellid.R
\name{trainLogit}
\alias{trainLogit}
\title{trainLogit}
\usage{
trainLogit(balanced, ...)
}
\arguments{
\item{balanced}{Balanced groups}

\item{...}{Extra param for .createFormula()}
}
\description{
Trains a logistic regression
}
